Ignore key during setup
Since the key will be the basis for encryption, we do not want it
committed to the repository.

If the repository has an `.gitignore` file, we simply append the file
with `invisible_ink.key`. Otherwise, we create the `.gitignore` and add
`invisible_ink.key`.

It should be noted that the presence of a `.gitignore` file does not
mean git is initialized in the repository.