Print decrypted file contents to console
Introduce `ActiveSupport::EncryptedFile`, and add `active_support` as a
dependency since this is the mechanism in which we want to encrypt and
decrypt files with.

Rather than require `active_support` in both `spec/spec_helper.rb` and
`exe/invisible_ink`, we instead add it to `lib/invisible_ink.rb`, along
with `securerandom`. That way, we can just require `../lib/invisible_ink` in
the executable.

Note that we're using `require_relative` and not `require`. This is
because, to the best of my knowledge, using `require_relative` will
ensure we're loading the correct version of the gem, versus whatever is
installed on the system. I think this would only matter during local
development.

We also introduce several helpers to simplify the tests.