Encrypt files when calling "write" command
This is the heart and soul of the gem. When running the `write` command,
we open a temporary file in the system `EDITOR`, allowing the user to
enter the desired content. Once the user saves and exists, the temporary
file is deleted, and the contents are encrypted.

If the file does not yet exist, we need to first create it via `write`,
but we pass `nil` so that it's empty. This is because we can't call
`change` unless the file exists.

We also need to ensure that the system `EDITOR` is set. If it's not, we
return a helpful error message. This was [borrowed from Rails][3].

Finally, we print a simple message if the process if interrupted when
editing the file. Again, this was [borrowed from Rails][4].

This is a hard feature to test, since it requires user input. I borrowed
from the [existing Rails test][1] from which this is based. It sets the
`ENV['EDITOR']` variable to a Ruby script that is then evaluated inline.

In order for this to work, we needed to introduce a test helper that is
responsible for temporarily switching environment variables. Again, this
was [borrowed from Rails][2].

To effectively test adding no content to the file, we default to using
the `cat` editor. This ensures no editor actually opens during the test
execution (which would happen if we chose `vim`, for example).

[1]: https://github.com/rails/rails/blob/e23e1718ead861ae43004f5706ddad780a440359/railties/test/commands/credentials_test.rb#L349
[2]: https://github.com/rails/rails/blob/e23e1718ead861ae43004f5706ddad780a440359/railties/test/env_helpers.rb#L21-L26
[3]:https://github.com/rails/rails/blob/05c3f84a511663cfef265834efb7ad8d3e472682/railties/lib/rails/command/helpers/editor.rb#L12-L21
[4]: https://github.com/rails/rails/blob/05c3f84a511663cfef265834efb7ad8d3e472682/railties/lib/rails/command/helpers/editor.rb#L30-L32