Handle cases where the key is missing
If the `invisible_ink.key` or `ENV["INVISIBLE_INK_KEY]` are missing,
`ActiveSupport::EncryptedFile::MissingKeyError` is raised. This is
because we set `raise_if_missing_key: true`.

In an effort to make the error render more clearly to the screen, we
rescue from that error only to `put` it to the screen, along with a hint
to run the setup script.